requirements.txt
pillow
pytesseract
pdf2image
pdfplumber
opencv-python-headless
PyMuPDF
rapidfuzz==3.6.1
streamlit==1.25.0
pandas
python-multipart
